+++ plugins/wp-chargify/includes/forms/account-details.php
@@ -1,0 +1,1 @@
+<?php
